Add files via upload
This was an extra credit assignment that's worth 5 points. I was bored, and wanted to show the teacher that I'm interested in the subject; So I wrote this script. I haven't coded in a while, so what I put into play is what I know. It was a great time, wrote this while watching Brian stream. Good times.